Bug in installing interrupt handlers for types

+++ src/booter/interrupts.c
@@ -226,8 +226,8 @@ void install_interrupt_handler(int num, void *handler) {
      *        INCORPORATE THE ABOVE COMMENTS IF YOU WISH.)
      */
 
-     interrupt_descriptor_table[num].offset_15_0 = (handler << 16) >> 16;
-     interrupt_descriptor_table[num].offset_31_16 = handler >> 16;
+     interrupt_descriptor_table[num].offset_15_0 = ((int) handler << 16) >> 16;
+     interrupt_descriptor_table[num].offset_31_16 = (int) handler >> 16;
      interrupt_descriptor_table[num].selector = SEL_CODESEG;
      interrupt_descriptor_table[num].type_attr = 0;
 }
